requirements.txt
Django
djangorestframework
mysqlclient
djangorestframework-simplejwt
drf-spectacular
python-dotenv